requirements.txt
gate-api==7.1.7
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
requests==2.31.0

